+++ python/lsst/ts/m1m3/utils/m1m3_bump_tests_times.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 # This file is part of ts_m1m3_utils
 #
 # Developed for the LSST Telescope and Site.
@@ -26,10 +24,15 @@ import asyncio
 import logging
 from urllib.parse import urlencode, urlunparse
 
+import sty
 from astropy.time import Time, TimeDelta
-from lsst.ts.m1m3.utils import BumpTestTimes, ForceActuatorForces
-from lsst.ts.xml.tables.m1m3 import FATable, ForceActuatorData, force_actuator_from_id
+from lsst.ts.xml.enums.MTM1M3 import BumpTest as BumpTestStatus
+from lsst.ts.xml.tables.m1m3 import FATable, force_actuator_from_id
 from lsst_efd_client import EfdClient
+
+from .bump_test_times import BumpTest, BumpTestTimes
+from .duration_time import DurationTime
+from .force_actuator_forces import ForceActuatorForces
 
 
 def parse_arguments() -> argparse.Namespace:
@@ -40,14 +43,14 @@ def parse_arguments() -> argparse.Namespace:
     parser = argparse.ArgumentParser(description="Queries bump test status.")
     parser.add_argument(
         "start_time",
-        type=Time,
+        type=DurationTime(now),
         default=now - TimeDelta(7, format="jd"),
         nargs="?",
         help="Start time in a valid format: 'YYYY-MM-DD HH:MM:SSZ'",
     )
     parser.add_argument(
         "end_time",
-        type=Time,
+        type=DurationTime(now),
         default=now,
         nargs="?",
         help="End time in a valid format: 'YYYY-MM-DD HH:MM:SSZ'",
@@ -81,6 +84,8 @@ def parse_arguments() -> argparse.Namespace:
 async def run_loop() -> None:
     args = parse_arguments()
 
+    start_t, end_t = DurationTime.pair(args.start_time, args.end_time)
+
     level = logging.INFO
 
     if args.d:
@@ -100,20 +105,22 @@ async def run_loop() -> None:
     for aid in [int(a) for a in args.actuators]:
         actuator = force_actuator_from_id(aid)
         logging.info(f"** Actuator {aid} type: {actuator.actuator_type}")
-        primary, secondary = await btt.find_times(aid, args.start_time, args.end_time)
 
-        async def print_bump(start: Time, end: Time) -> None:
-            def act(index: int | None, actuator: ForceActuatorData) -> int:
-                return 0 if index is None else actuator
-
+        async def print_bump(test: BumpTest) -> None:
             params = {
                 "refresh": "Paused",
-                "tempVars[x_index]": act(actuator.x_index, actuator.actuator_id),
-                "tempVars[y_index]": act(actuator.y_index, actuator.actuator_id),
-                "tempVars[z_index]": actuator.actuator_id,
-                "tempVars[s_index]": act(actuator.s_index, actuator.actuator_id),
-                "lower": start.isot + "Z",
-                "upper": end.isot + "Z",
+                "tempVars[x_index]": (
+                    0 if test.fa.x_index is None else test.fa.actuator_id
+                ),
+                "tempVars[y_index]": (
+                    0 if test.fa.y_index is None else test.fa.actuator_id
+                ),
+                "tempVars[z_index]": test.fa.actuator_id,
+                "tempVars[s_index]": (
+                    0 if test.fa.s_index is None else test.fa.actuator_id
+                ),
+                "lower": test.start_time.isot + "Z",
+                "upper": test.end_time.isot + "Z",
             }
             url = urlunparse(
                 (
@@ -133,9 +140,16 @@ async def run_loop() -> None:
                     "",
                 )
             )
-            print(start.isot, end.isot, url)
+            print(
+                sty.fg.green if test.result == BumpTestStatus.PASSED else sty.fg.red,
+                test.start_time.isot,
+                test.end_time.isot,
+                test.result,
+                sty.fg.rs,
+                url,
+            )
             if args.details:
-                faf = ForceActuatorForces(start, end, client)
+                faf = ForceActuatorForces(test.start_time, test.end_time, client)
                 fa_fe = await faf.actuator_following_error(actuator)
                 print(
                     f"Following errors min: {fa_fe.primary.min():.3f} N "
@@ -148,16 +162,25 @@ async def run_loop() -> None:
                     f"max {flat_fe.max():.3f} N"
                 )
 
-        print("Primary bump tests")
-        for bump in primary:
-            await print_bump(bump[0], bump[1])
+        print(sty.fg.yellow, "Primary bump tests - FA", actuator.actuator_id, sty.bg.rs)
+        async for bump in btt.find_times(actuator, True, start_t, end_t):
+            await print_bump(bump)
 
-        print("===================")
-        print("Secondary bump tests")
-        for bump in secondary:
-            await print_bump(bump[0], bump[1])
+        if actuator.s_index is not None:
+            print(sty.bg.blue, "\u25A9" * 50, sty.bg.rs)
+            print(
+                sty.fg.yellow,
+                "Secondary bump tests - FA",
+                actuator.actuator_id,
+                sty.fg.rs,
+            )
+            async for bump in btt.find_times(actuator, False, start_t, end_t):
+                await print_bump(bump)
 
-    await client._influx_client.close()
+    if client.influx_client is None:
+        await client._influx_client.close()
+    else:
+        await client.influx_client.close()
 
 
 def run() -> None: